Updated a script for bumping beta release version
