+++ .github/workflows/create-release-branch.yml
@@ -129,13 +129,3 @@ jobs:
           echo "Creating pull requests by GitHub Action is no longer supported per Microsoft Open Source Security Policy."
           echo "Please create a pull request manually from ${{ github.event.inputs.branch }} to main."
           echo "See example: https://github.com/Azure/communication-ui-library/pull/2631"
-
-      # if a stable release bump the package versions to beta.0 numbers for next beta release
-      - name: Bump versions for pre-release branch if stable release
-        if: ${{ startsWith(github.event.inputs.branch, 'prerelease-stable') }}
-        run: |
-          node common/scripts/bump-beta-version.js
-          rush update
-          git add .
-          git commit -m "Bump package versions to beta.0 for next release"
-          git push
