+++ common/scripts/bump-beta-release-version.js
@@ -1,0 +1,45 @@
+const path = require('path');
+const { updateAllVersions, findAllPackageJSON, getAllNames, updateAllDepVersions } = require('./package-utils');
+
+const PACKAGES_DIR = path.join(__dirname, '..', '..', 'packages');
+
+const bumpType = process.argv[2];
+
+const main = () => {
+  const packagePaths = findAllPackageJSON(PACKAGES_DIR);
+  const depNames = getAllNames(packagePaths);
+  if (!['major', 'minor', 'patch', 'beta'].includes(bumpType)) {
+    throw '\nplease add either major/minor/patch/beta as a parameter!\n\n  Syntax:\n  node bump-beta-release-version.js minor\n'
+  }
+  updateAllVersions(bumpBetaVersion);
+  // Need to update all internal project dependencies using the same rule
+  updateAllDepVersions(bumpBetaVersion, depNames);
+}
+
+// https://skype.visualstudio.com/SPOOL/_wiki/wikis/SPOOL.wiki/43834/How-to-determine-the-next-web-UI-Library-package-version.
+const bumpBetaVersion = (currentVersion) => {
+  const versionStrs = currentVersion.split('-beta.');
+  const [major, minor, patch] = versionStrs[0].split('.');
+
+  if (bumpType === 'beta') {
+    const betaVersion = versionStrs[1];
+    // if bumpType is beta, no need to update major, minor or patch version
+    if (betaVersion === undefined) {
+      throw '\nplease use either major/minor/patch as a parameter!\n\n  Syntax:\n  node bump-beta-release-version.js minor\n'
+    }
+    const newBeta = Number.parseInt(betaVersion) + 1;
+    return `${newMajor}.${newMinor}.${newPatch}-beta.${newBeta}`
+  } else {
+    const newMajor = bumpType === 'major' ? Number.parseInt(major) + 1 : major;
+    const newMinor = bumpType === 'major' ?
+      0
+      : bumpType === 'minor' ? Number.parseInt(minor) + 1 : minor;
+    const newPatch = bumpType === 'major' || bumpType === 'minor' ?
+      0
+      : bumpType === 'patch' ? Number.parseInt(patch) + 1 : patch;
+    
+    return `${newMajor}.${newMinor}.${newPatch}-beta.1`
+  }
+}
+
+main();
